lgshop/requirements.txt
amqp==5.0.5
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
Django==2.2
django-haystack==3.0
django-redis==4.12.1
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py @ file:///D:/%E5%95%86%E5%9F%8E%E7%B3%BB%E7%BB%9F%E6%9D%90%E6%96%99/%E4%B8%8A%E8%AF%BE%E8%B5%84%E6%96%99/%E4%B8%8A%E8%AF%BE%E8%B5%84%E6%96%99/fdfs_client-py-master.zip
idna==2.10
importlib-metadata==3.7.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
kombu==5.0.2
mutagen==1.45.1
mysqlclient==2.0.3
openpyxl==3.0.7
paramiko==2.7.2
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
pycparser==2.20
pycryptodomex==3.9.4
Pygments==2.8.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
pytz==2021.1
QQLoginTool==0.3.0
redis==3.5.3
requests==2.25.1
ronglian-sms-sdk==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
zipp==3.4.1
